Split out BaseRuntime
